Added common responsemodel
added in common api response model

+++ Controllers/EmployeeController.cs
@@ -1,4 +1,5 @@
-﻿using EmployeeApi.Models;
+﻿using EmployeeApi.Common;
+using EmployeeApi.Models;
 using EmployeeApi.Repository;
 using EmployeeApi.Services;
 using Microsoft.AspNetCore.Authorization;
@@ -12,7 +13,6 @@ namespace EmployeeApi.Controllers
     [Route("api/[controller]")]
     public class EmployeeController : ControllerBase
     {
-        //private readonly EmployeeRepository _repo;
         private readonly IEmployeeService _employeeService;
 
         public EmployeeController( IEmployeeService employeeService)
@@ -23,7 +23,12 @@ namespace EmployeeApi.Controllers
         [HttpGet]
         public IActionResult GetAll()
         {
-            return Ok(_employeeService.GetAll());
+            return Ok(new ApiResponse<List<Employee>>
+            {
+                Success = true,
+                Message = "Data retrieved successfully",
+                Data = _employeeService.GetAll()
+            });
         }
 
         [HttpGet("{id}")]
@@ -31,8 +36,21 @@ namespace EmployeeApi.Controllers
         {
             var emp= _employeeService.GetById(id);
             if(emp == null)
-                return NotFound();
-            return Ok(emp);
+            {
+                return NotFound(new ApiResponse<string>
+                {
+                    Success = false,
+                    Message = "Employee not found",
+                    Data = null
+                });
+            }
+                
+            return Ok(new ApiResponse<Employee>
+            {
+                Success = true,
+                Message = "Retrieved successfully",
+                Data = emp
+            });
         }
 
         [Authorize(Roles = "Admin")]
@@ -40,21 +58,59 @@ namespace EmployeeApi.Controllers
         public IActionResult AddEmployee(Employee employee)
         {
             _employeeService.Add(employee);
-            return Ok();
+            return Ok(new ApiResponse<Employee>
+            {
+                Success = true,
+                Message = "Employee added successfully",
+                Data = employee
+            });
         }
 
         [HttpPut]
         public IActionResult UpdateEmployee(Employee employee)
         {
+            var existing = _employeeService.GetById(employee.Id);
+
+            if (existing == null)
+            {
+                return NotFound(new ApiResponse<string>
+                {
+                    Success = false,
+                    Message = "Employee not found",
+                    Data = null
+                });
+            }
             _employeeService.Update(employee);
-            return Ok();
+
+            return Ok(new ApiResponse<Employee>
+            {
+                Success = true,
+                Message = "Employee updated successfully",
+                Data = employee
+            });
         }
         
         [HttpDelete("{id}")]
         public IActionResult DeleteById(int id)
         {
+            var existing = _employeeService.GetById(id);
+
+            if (existing == null)
+            {
+                return NotFound(new ApiResponse<string>
+                {
+                    Success = false,
+                    Message = "Employee not found",
+                    Data = null
+                });
+            }
             _employeeService.Delete(id);
-            return Ok();
+            return Ok(new ApiResponse<string>
+            {
+                Success = true,
+                Message = "Employee deleted successfully",
+                Data=null
+            });
         }
     }
 }
